// src/DockiUp.Frontend/src/app/api/model/imageUpdateDto.ts
/**
 * DockiUp.API | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ImageUpdateDto {
    projectId: string;
    serviceName: string;
    image: string;
    currentDigest: string | null;
    latestDigest: string | null;
    checkedAt: string;
    updateAvailable: boolean;
    note: string | null;
}