// src/DockiUp.Frontend/src/app/api/model/imageUpdatePolicy.ts
/**
 * DockiUp.API | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ImageUpdatePolicy = 'Off' | 'Notify' | 'Auto';

export const ImageUpdatePolicy = {

    Off: 'Off' as ImageUpdatePolicy,

    Notify: 'Notify' as ImageUpdatePolicy,

    Auto: 'Auto' as ImageUpdatePolicy
};