// src/DockiUp.Frontend/src/app/api/model/imageUpdateSettingsDto.ts
/**
 * DockiUp.API | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ImageUpdatePolicy } from './imageUpdatePolicy';


export interface ImageUpdateSettingsDto {
    policy: ImageUpdatePolicy;
    excludedServices: Array<string>;
    pinnedImages: { [key: string]: string; };
}
export namespace ImageUpdateSettingsDto {
}